+++ nbrmd/cli.py
@@ -9,51 +9,7 @@ from nbrmd import readf, writef
 from nbrmd import writes
 from .languages import get_default_language
 from .combine import combine_inputs_with_outputs
-
-
-def convert_nbrmd(nb_files, in_place=True, combine=True):
-    """
-    Export R markdown notebooks, or Jupyter notebooks, to the opposite format
-    :param nb_files: one or more notebooks
-    :param in_place: should result of conversion be stored in file
-    with opposite extension?
-    :param combine: should the current outputs of .ipynb file be preserved,
-    when a cell with corresponding input is found in .Rmd?
-    :return:
-    """
-    for nb_file in nb_files:
-        file, ext = os.path.splitext(nb_file)
-        if ext not in ['.ipynb', '.Rmd']:
-            raise TypeError(
-                'File {} is neither a Jupyter (.ipynb) nor a '
-                'R Markdown (.Rmd) notebook'.format(nb_file))
-
-        nb = readf(nb_file)
-
-        if in_place:
-            if ext == '.ipynb':
-                nb_dest = file + '.Rmd'
-                print('Jupyter notebook {} being converted to '
-                      'R Markdown {}'.format(nb_file, nb_dest))
-            else:
-                msg = ''
-                nb_dest = file + '.ipynb'
-                if combine and os.path.isfile(nb_dest):
-                    try:
-                        nb_outputs = readf(nb_dest)
-                        combine_inputs_with_outputs(nb, nb_outputs)
-                        msg = '(outputs were preserved)'
-                    except (IOError, NotJSONError) as err:
-                        msg = '(outputs were not preserved: {})'.format(err)
-                print('R Markdown {} being converted to '
-                      'Jupyter notebook {} {}'
-                      .format(nb_file, nb_dest, msg))
-            writef(nb, nb_dest)
-        else:
-            if ext == '.ipynb':
-                print(writes(nb))
-            else:
-                print(ipynb_writes(nb))
+from .file_format_version import check_file_version
 
 
 def cli_nbrmd(args=None):
@@ -83,48 +39,60 @@ def nbrmd():
     :return:
     """
     args = cli_nbrmd()
-    convert_nbrmd(args.notebooks, args.in_place, args.preserve_outputs)
+    convert(args.notebooks, args.in_place, args.preserve_outputs, True)
 
 
-def convert_nbsrc(nb_files, in_place=True, combine=True):
+def convert(nb_files, in_place=True, combine=True, markdown=False):
     """
-    Export python or R scripts, or Jupyter notebooks, to the opposite format
+    Export R markdown notebooks, python or R scripts, or Jupyter notebooks,
+    to the opposite format
     :param nb_files: one or more notebooks
+    :param markdown: R markdown to Jupyter, or scripts to Jupyter?
     :param in_place: should result of conversion be stored in file
     with opposite extension?
     :param combine: should the current outputs of .ipynb file be preserved,
-    when a cell with corresponding input is found in .py or .R file?
+    when a cell with corresponding input is found in .Rmd/.py or .R file?
     :return:
     """
     for nb_file in nb_files:
         file, ext = os.path.splitext(nb_file)
-        if ext not in ['.ipynb', '.py', '.R']:
-            raise TypeError(
-                'File {} is neither a Jupyter (.ipynb) nor a '
-                'python script (.py), nor a R script (.R)'.format(nb_file))
+        if markdown:
+            format = 'R Markdown'
+            if ext not in ['.ipynb', '.Rmd']:
+                raise TypeError(
+                    'File {} is neither a Jupyter (.ipynb) nor a '
+                    'R Markdown (.Rmd) notebook'.format(nb_file))
+        else:
+            format = 'source'
+            if ext not in ['.ipynb', '.py', '.R']:
+                raise TypeError(
+                    'File {} is neither a Jupyter (.ipynb) nor a '
+                    'python script (.py), nor a R script (.R)'.format(nb_file))
 
         nb = readf(nb_file)
         main_language = get_default_language(nb)
-        ext_dest = '.R' if main_language == 'R' else '.py'
+        ext_dest = '.Rmd' if markdown else '.R' \
+            if main_language == 'R' else '.py'
 
         if in_place:
             if ext == '.ipynb':
                 nb_dest = file + ext_dest
                 print('Jupyter notebook {} being converted to '
-                      'source {}'.format(nb_file, nb_dest))
+                      '{} {}'.format(nb_file, format, nb_dest))
             else:
                 msg = ''
                 nb_dest = file + '.ipynb'
                 if combine and os.path.isfile(nb_dest):
+                    check_file_version(nb, nb_file, nb_dest)
                     try:
                         nb_outputs = readf(nb_dest)
                         combine_inputs_with_outputs(nb, nb_outputs)
                         msg = '(outputs were preserved)'
                     except (IOError, NotJSONError) as error:
                         msg = '(outputs were not preserved: {})'.format(error)
-                print('R Markdown {} being converted to '
+                print('{} {} being converted to '
                       'Jupyter notebook {} {}'
-                      .format(nb_file, nb_dest, msg))
+                      .format(format, nb_file, nb_dest, msg))
             writef(nb, nb_dest)
         else:
             if ext == '.ipynb':
@@ -160,4 +128,4 @@ def nbsrc():
     :return:
     """
     args = cli_nbsrc()
-    convert_nbsrc(args.notebooks, args.in_place, args.preserve_outputs)
+    convert(args.notebooks, args.in_place, args.preserve_outputs, False)