+++ notebook.nb.py
@@ -1,0 +1,7 @@
+---
+jupyter:
+  nbrmd_formats: ipynb,py
+  nbrmd_file_format_version: '1.0'
+---
+
+# New cell