+++ tests/test_ipynb_to_rmd.py
@@ -5,6 +5,8 @@ import nbrmd
 from .utils import list_all_notebooks, remove_outputs, \
     remove_outputs_and_header
 
+nbrmd.file_format_version.FILE_FORMAT_VERSION = {}
+
 
 @pytest.mark.parametrize('nb_file', list_all_notebooks('.ipynb'))
 def test_identity_source_write_read(nb_file):