+++ tests/test_trust_notebook.py
@@ -1,8 +1,11 @@
 import os
 import shutil
 import pytest
+import nbrmd
 from nbrmd.contentsmanager import RmdFileContentsManager
 from .utils import list_all_notebooks
+
+nbrmd.file_format_version.FILE_FORMAT_VERSION = {}
 
 
 @pytest.mark.parametrize('nb_file', list_all_notebooks('.py'))
@@ -41,8 +44,7 @@ def test_ipynb_notebooks_can_be_trusted(nb_file, tmpdir):
     # Sign notebook explicitely (save it, and reload without
     # validating to remove 'trusted' metadata in cells)
     cm.save(nb, py_file)
-    nb = cm._read_notebook(tmp_py)
-    cm.notary.sign(nb)
+    cm.trust_notebook(py_file)
 
     nb = cm.get(file)
     for cell in nb['content'].cells: